CL node keys are not displayed in the node ops UI when a key exists that isn't part of a configured chain (#8583)
* Evaluate many different ways to fix/modify this - ultimately went with the simplest and for once, kicking the can down the road is actually the right thing to do.
This is a fix for node side configuration state that is NOT considered an error, but the operator-ui is built to assume would be an error. Eth Keys in storage that are not part of a currently configured chain are not considered orphaned although they may seem that way.

* Our chain support has gotten slightly out of sync. So add support for StarkNet.
Added an extra error handling case if we get out of sync again.
Move to using package defined const/enum in place of the original direct string comparisons now that they are available.
Updated the enums in the graphql so the operator UI validation can be up to date as well.
Updated a test that was previously meant to trigger an error on code that now no longer returns an error if a key does not belong to an active chain.
Resolver/graphql changes were tested against the operator UI including yarn tests

* convert comment block

* only include keys with valid chains

---------

Co-authored-by: Sam Hassan <es.hassan187@gmail.com>