Restore rename to OnDeleteJob, restore legacy BeforeJobDeleted for webhook
There's no easy way to solve this for webhook.  They need to be able
load the webhook specs, to know what to send to the EI's, but that's
already been deleted from the db by the time OnDeleteJob() is called.

The only alternative would be to have the spawner make a special
exception for webhook jobs, loading the webhook specs into the
job.Job struct manually, and passing it along with the spec to
BeforeDeleteJob().  (For generic jobs, it just callas orm.FindJob()
which doesn't populate it with the webhook specs.)