Remove support for arbitrary client cancellation of replays
We were sharing two different contexts between two different goroutines,
lp.Replay() and lp.run(), passing one context across a channel, and
storing the other in a struct.

In addition to being outside of standard golang practice, using
the replay context in lp.run() during replay meant that any shutdown
event which occurs during a replay would result in a non-graceful
shutdown.  The only simple way to fix the shutdown bug would be to
have lp.run() respond immediately if either context is cancelled;
however, this kind of context merging goes further outside of golang
paradigm.

This commit cleans up the context handling by removing all references
to lp.ctx from lp.Replay() and all references to replayCtx from lp.run(),
so that they each have clear ownership of only a single context and only
respond to their own context being cancelled.

The only downside of this is it removes the ability of callers of
lp.Replay() to interrupt lp.run() after it's begun its replay of
logs.  However, this is at present an unnecessary feature; we don't
currently use it for anything other than shutdown events... which are
already handled by lp.run().  If we do need such a feature in the
future, we can add a cancelReplay channel for it, possibly combining
that with the startReplay and/or completeReplay channels.