Convert TestLogPoller_DBErrorHandling to blackbox
This test needed to be updated anyway, since it used lp.done.  While doing so, I
noticed that the only reason it was a whitebox test in the first place was so that
lp.ctx could be initialized before calling lp.Replay() without requiring
the extra  overhead of starting the entire logpoller with lp.Start()
(then stopping it later and restarting it).  However, with this PR lp.Replay()
will no longer depend on lp.ctx so the test can be fully black box.