Split plugin service into multiple instances
The primary focus of splitting into multiple instances is taking slices of the
total number of upkeep ids per instance such that there is no overlap. This
involves a change to the registry implementation to take the mod of an upkeep
id and determine if the remainder equals the instance id.

Each instance of the plugin requires a different registry instance. Futher
cleanup can be done to reduce resource usage, but the general idea should
function correctly.