Improve robustness of TestLogPollerger_Replay tests
The main problem was a missing lp.replayComplete<- , but I've also
increased some timeouts, and added an extra 4s failsafe timeout around
lp.Wait

The last timeout is to avoid the whole test from waiting the full 10
mins to fail, in the event that one or more goroutines inside logpoller
get stuck (eg. waiting on <-lp.replayComplete, as happened here) and
it never proceeds past lp.Wait to the next test.  This shouldn't be able
to happen now, but it should  make it easier to debug if anything similar
comes up.

There is also a rare race condition inside testify that can happen when
lp.ctx is passed as an argument to mock.On().  Replaced this with
mock.Anything and removed the first mock.On().Maybe() which was the
only reason it was needed.  This appears to be unnecessary based on
local testing with various -parallel and -p settings.