Fix race condition in TestLogPoller_Replay/client_abort_doesnt_hang_run_loop
This fixes another race condition in this test.

The second call to FilterLogs() happens while the replay is ongoing.
If it returns before cancel() is called then there is a chance that the
whole replay can complete successfully.  Cancelling first should
guarantee that it can't complete succesfully--next db call will fail
due to cancelled context.