fix: log in goroutine after test ends
The persistence layer of ocr2vrf spawns a goroutine to store the DKG
shares, unfortunately it causes flakes in the integration tests. Making
the shutdown clean is not a trivial task, so we're just going to move
the logs to ocr2vrf instead of having them in the goroutine itself.