Reintegrate dependencies
The flattened repo is now reintegrated with all supporting dependencies. One new
difference to note is the use of factory methods being provided to the delegate
instead of fully prepared instances. This is to support the plugin life-cycle
by requiring all dependencies also adhere to the same life-cycle.

Other improvements include better parallelization and component reuse across the
application.