Remove abandoned tx attempts before attempting INSERT
Removing the abandoned transaction after an error is detected doesn't work, because it poisons the whole transaction. Even though subsequent queries would have succeeded, they fail because there's already been a sql error.

This would be an ideal case for use of nested transactions, but we can't easily add support for that so the only way is to check for and remove any conflicting abandoned transactions before the INSERT. We can at least combine the check and DELETE into a single query.