Separate pass / exit-safety checks from each other
The wait groups were originally being used as an assert.Eventually(), to
wait for the condition when both calls to FilterLogs() have been
succefully made.  Then they were re-purposed to also safegaurd against a
subtest from exiting before all of its goroutines have exited.

This caused a fundamental conflict in where to place the wg.Add() calls;
to ensure the test doesn't fail immediately (before one or both of the calls to
FilterLogs() happens), they had to be placed outside of these calls.
However, that doesn't work for the second use case... because then if
they don't get called (either because the 2s timeout expired or due to some
problem in the code being tested), then the subtest is blocked from
exiting even though it hasn't spawned any goroutines.

We saw this happen in CI in Test_Replay/shutdown_during_replay
(presumably due to the 2s timeout getting hit?), but it looks like the same
could have happened in the previous subtest as well.

This splits pass & exit-safety logic from each other so they can each
behave correctly, and increases the timeouts from 2s to 10s