Implement OnCreateJob() & OnDeleteJob for ocrbootstrap delegate
Unlike the OCR2 job delegate, the OCRBootstrap delegate does not have
access to the ChainSet so it has no way of finding LogPoller to call
lp.Register() or lp.Unregister().  Instead, the relay itself must be
extended to be able to do this on demand.  RegisterFilters() and
UnregisterFilters() have been added for this purpose.