BCF-2368: ensure unique name per LOOPP (#9762)
End to end work uncovered a problem where a plugin would not start
because the prom server couldn't get a port. Ultimately this was caused
by duplicate names in the LoopRegistry.

This change makes names unique and, to ensure better error handling, it
changes the Register function to error if called more than once. The latter
change will make the problem much more clear should it ever happen again.