Fix flakey TestLogPoller_Integration test
This was failing due to unpredictability of when the backup log poller runs.  The backup log poller starts 100 blocks behind the regular log poller, so for tests this means it effectively always starts at block 0.  It runs the first time after 100ms delay, so depending on circumstances you may have some additional prior blocks before the fromBlock of the Replay.  These are harmless, but weren't accounted for by the tests.  It should be predictable now because it waits until blocks before fromBlock are there initially, to make sure that the backup poller has already run.  When it gets to looking for Log2 events (after the second replay) we can be sure it *won't* run again since the delay after that is set to 1 hour in tests.

Also:
- fixes a bug in the handling of a logpoller initialization edge case:
  If Replay() is called but the filters have not been successfully loaded yet, it will try once to load them.  If they haven't been loaded it most likely means that connectivity to the db has been down since the node started, and it's still retrying to load them in another thread.  The bug was that if it succeeds in loading them, it was skipping the replay instead of going ahead with it.  This should only matter if the db connectivity suddenly came up and started working right at the moment Replay is called, so we likely haven't run into it.

- Lowercases some error messages which were capitalized