Fix flaky integration test for log provider
The backfill integration test was relying on concurrency and timers but had
inconsistent results when run on the CI environment, likely to do resource
consumption. The solution was to add a new exposed function to call necessary
processes directly in the test and not require running multiple processes. This
reduced the test time from 12s to 7s and should be less flaky.