feature: heartbeat stores for bhs
Based on a block number job spec parameter, call Store() on the BHS
every `heartbeatPeriodBlocks` number of blocks.

This will ensure that we have an anchor for backwards mode in case we
ever run into an issue in prod.

Note that this is based upon the head broadcaster rather than the log
poller for redundancy reasons. We have seen the log poller unable to
track heads sometimes whereas the head tracker is still doing it just
fine.