fix: arbitrum chain id check in all CSU methods
The arbitrum chain id check needs to be the same in all of the methods
exported in the ChainSpecificUtil contract. Prior to this change
getBlockNumber, getCurrentTxL1GasFees and getL1CalldataGasCost did not
correctly check for all possible arbitrum chain ids.

To prevent this kind of error in the future, a new private function
isArbitrumChainId was created and used in all methods that need to check
the chain id.