Fix redundant "replacing abandoned tx" logs (#11019)
This fixes the issue of incorrectly logging the "replacing abandoned tx"
log. The DELETE query doesn't return a sql.ErrNoRows on failing to find
rows to be deleted. Instead we are checking on the count of RowsAffected.
Another minor fix is continuing to execute the function if no attempts
were deleted, instead of throwing an error, since at this point there is
no need to break the execution flow.

Co-authored-by: Simson <simsonraj.easvarasakthi@smartcontract.com>