Update SimulatedBackendClient CallContext (#11164)
* Update SimulatedBackendClient CallContext

The function `CallContext` has different supported contract function calls than
`BatchCallContext` even though the latter is simply a batch version of the
former.

This commit makes the two functions match both in the supported calls, but also
in the validation and execution of those calls.

* Update core/chains/evm/client/simulated_backend_client.go

Use suggestion on default address returned on error.

Co-authored-by: Jordan Krage <jmank88@gmail.com>

* replace errors with panics

* align parameters to documentation, real-world RPCs, and the simulated backend client

* allow different receipt types

* comment on strong typing on result

* remove logic for handling incomplete transaction receipt handling on simulated backend client

* reduce panics in favor of errors

* remove commented code

---------

Co-authored-by: Jordan Krage <jmank88@gmail.com>