fix: handle stale blockhashes
One problem we have been running into is that requests that get re-orged
on polygon sometimes do not end up having their log re-sent by the log
broadcaster. In the event this happens, we'd like to still be able to
correctly process the request.

This change adds a check where the log for the request is re-fetched
from the RPC node using FilterLogs, and the blockhash and block number
are compared to the block hash and the block number in the pending
request object. If they are different, we prefer the block hash from
the FilterLog call since that should eventually be the correct value.

In simulateFulfillment, we determine what FilterLogs says and we
potentially use that value in the simulation (if no errors are thrown
when calling FilterLogs). We also add new fields to vrfPipelineResult
called blockNumberUsed and blockHashUsed in case these are different
from the request block number and block hash, and are used in logging
for potential debugging purposes.