Feed LatestPrice query cache (#11326)
* Price retrieval cache for Mercury

Problem: Core nodes are querying LatestReport a lot (~70rps). This is not only wasteful, it potentially increases latency if Mercury server falls behind and is also completely unnecessary. We do not need such up-to-the-millisecond accuracy for billing,

Proposed solution: Introduce a price cache for mercury with a TTL e.g. of 1 minute or whatever tolerance we have for billing price and only fetch from the server when the cache becomes stale.

* fix

* Some PR comments

* Fix more PR comments

* PR comments

* lint