Change UseFinalityTags to FinalityTagEnabled in CHANGELOG.md to avoid confusion
The name of the param LogPoller accepts is useFinalityTag, but the externally-facing name of the toml config param it's set based on is FinalityTagEnabled. There is also a toml config param named UseFinalityTag but it's only used for integration tests, there is no such thing in the main production toml config. Also, adding EVM prefix for clarity.