Handle edge case involving blocks not being found in the db (#11298)
If there is a backfill followed by an error, logs get written to the
db but no blocks. This will make the logpoller (or backup log poller)
rely on the chain next time instead of the db to determine
lastFinalizedBlock, which could result in a gap in logs processed.
Fixing this by writing the last block in the backfil to the db along
with the logs. Its lastFinalizedBlock field will be set to its own
block number + 1, so if the db crashes it should start by pulling
the logs from that one.