feat: add liquidity balancer specs table (#11564)
Add the liquidity balancer specs table. This table will be used to house
the specs for the liquidity balancer job which will be implemented in
ccip rather than core.

The entirety of the spec's content will be in the
`liquidity_balancer_config` field. This is done to minimize future
migrations. This doesn't mean we won't be doing strict checks on the
contents of the config during the validation process.