chore(ccipdata): split ExecOffchainConfig definitions (#405)
## Motivation

Currently, we use the `ExecOffchainConfig` struct in two contexts:
* As the expected configuration format that we decode from the plugin configuration.
* As an internal interface type between the offramp reader interface and the execution reporting plugin.

So far, that has worked well, but this coupling hinders our ability to evolve the configuration format in v1.3 without breaking compatibility with older versions.

## Solution

This change splits the `ExecOffchainConfig` struct into two types:
 * A version-dependent struct we parse from JSON (defined in `v1_0_0/offramp.go`)
 * A version-independent interface type defined in offramp_reader.go.

It also documents the fields in the structs and adds a test for the config struct validation.

This change has no functional changes. A future change will implement a v1.3-specific offramp reader using the new configuration struct.