[lock/unlock]: run rebalancer locally (#469)
## Motivation
We want to be able to run the rebalancer DON locally so that we can
quickly iterate on features and fix bugs.

## Solution

Change Summary:
* add a `receive() payable` function to Rebalancer.sol so that it can
accept native token. The Rebalancer needs to be funded with native in
order to transfer native to the bridge adapter, which then transfers it
to the native bridge.
* add a core/scripts/ccip/rebalancer directory to house rebalancer
deployment and configuration scripts. We can eventually move this to
ccip-scripts but this is faster for now. See the Lock/Unlock space on
notion to see how to run the rebalancer don locally.
* thread safety for liquidity graph and liquidity manager registry
updates.
* pass in a logger to the EVM rebalancer

---------

Co-authored-by: Rens Rooimans <github@rensrooimans.nl>
Co-authored-by: dimitris <dimitrios.kouveris@smartcontract.com>