Token pool refactor (#453)
## Motivation
Currently, token pools have to allow each individual onRamp and offRamp.
This means that when a lane gets upgraded, many new approvals need to be
given. With this new change, pools simply allowList a chain, and get the
correct ramp information via the Router. This gives an additional
security benefit of the router being the source of truth for supported
lanes, and any removal from the router will cut off the old lane from
token pool access as well.