[lock/unlock] feat: arb L1<->L2 bridge plugin support (#490)
This PR implements off-chain bridging support for Arbitrum and for the
Mock Bridge being used in the integration tests. In order to properly
test this logic in the integration tests a lot of changes unfortunately
had to be made and this PR is quite big. This description aims to
summarize the changes on each part of the rebalancer codebase so far.

## Bridges
"Bridging support" in the context of the rebalancer plugin involves
mainly two things as of time of writing (and may potentially involve
more things in the future as we develop this plugin):

1. Determining transfer state: The plugin needs to know the state of
each pending transfer that was initiated using a bridge. Certain states
allow the plugin to perform certain actions. For example, a pending
transfer in the state "finalized" means that the plugin can instruct the
Rebalancer contract onchain to inject liquidity into the attached
liquidity container.
2. Determining bridge-specific arguments: Each "side" of the bridge
involves using different sets of arguments for many bridges, and
therefore bridges are directional - e.g L1 -> L2 and L2 -> L1 are
different bridge implementations. For example, for the L2 -> L1
direction, the bridge-specific arguments determined for finalizing
withdrawals on L1 are fundamentally different than finalizing transfers
in the L1 -> L2 direction.

The `Bridge` interface in
`core/services/ocr2/plugins/rebalancer/bridge/bridge.go` is the
interface that will be implemented for all bridge directions and chains
supported. See the interface for the exact methods, which have
descriptive comments.

### Arbitrum L1 <-> L2 Bridge Implementations

In this PR, the `Bridge` interface referenced above is implemented for
both directions of the Arbitrum One bridge, L1 -> L2 and L2 -> L1. Their
contents are mostly based off of the scripts merged in #480, so in
essence there is not much new here other than rejigging some things in
order to be able to properly implement the logic housed in those
scripts.

## Graphs and Discovery

Discovery of the graph of liquidities was tightly coupled to the EVM
implementation of the `Rebalancer` interface in the `liquiditymanager`
package, and this was causing issues for initializing the bridges, since
they also need to know a lot of information about the graph as well,
mainly the rebalancer and bridge adapter contract addresses.

As such, this PR factors out the discovery from `liquiditymanager` into
a new `discoverer` package. This package provides the `Discoverer`
interface which is able to discover rebalancers starting from a
master/root/main rebalancer contract and network.

Right now only an EVM implementation is provided, which will be
sufficient for our needs, but can be expanded in the future if there are
non-EVM rebalancers.

The `liquiditygraph` package was renamed to `graph` since it is now
housing much more than just liquidity data:

```golang
type XChainRebalancerData struct {
	RemoteRebalancerAddress   models.Address
	LocalBridgeAdapterAddress models.Address
	RemoteTokenAddress        models.Address
}

type Data struct {
	Liquidity         *big.Int
	TokenAddress      models.Address
	RebalancerAddress models.Address
	XChainRebalancers map[models.NetworkSelector]XChainRebalancerData
}
```

The data housed now closely mirrors what is on-chain, since pretty much
all of. this is needed in the off-chain components (e.g bridge
implementations).

## Rebalancing Algorithms

The `random` and `dummy` algorithms have been deleted to ease
maintenance burdens and only `pingpong` is now available and used within
tests.

## Plugin Logic

Plugin logic has been changed in order to account for fetching
bridge-specific data. The phases now operate as follows:

1. `Observation`: the plugin syncs the rebalancer graph, by doing a
discovery and updating the liquidities on each rebalancer. This behavior
was there before and is unchanged. *Additional behavior*: we check the
previous outcome and check if there are any _proposed transfers_ from
the committed outcome. These outcomes are then _resolved_ by using the
bridge implementations. See `resolveProposedTransfers` in
`core/services/ocr2/plugins/rebalancer/plugin.go` for more details.
2. `Outcome`: the plugin constructs an outcome by: (a) computing median
liquidity per-chain (b) computing the consensus graph of connections
between chains (c) consensus on pending transfers. This behavior was
there before and is unchanged. *Additional behavior*: we compute a
quorum on the _resolved transfers_ that have been sent as part of the
observation. This is similar to the pending transfers consensus but
involves a little more details. See `computeResolvedTransfersQuorum` in
`core/services/ocr2/plugins/rebalancer/plugin.go` for details.
3. `Reports`: the previous plugin logic here was actually buggy. In
order to make it clearer what this function is supposed to do the code
now determines _incoming_ and _outgoing_ transfers for each network.
_Incoming_ transfers are transfers that are meant to generate `Receive`
instructions in the report, and _outgoing_ transfers are transfers that
are meant to generate `Send` instructions in the report. See `Reports`
in `core/services/ocr2/plugins/rebalancer/plugin.go` for more details on
this.
4. `ShouldTransmitAcceptedReport`: check the sequence number that has
been transmitted onchain and compare it to the provided `seqNr`. If
`onchainSeqNr < seqNrArg` then we return `true`, otherwise we return
`false`.

---------

Co-authored-by: dimitris <dimitrios.kouveris@smartcontract.com>