offchain - batch caller verbose errors and token pools fix (#533)
## Motivation

- 1.0.0 Token pools do not have a `typeAndVersion` method which leaded
to rpc call failures.

- Also in this scenario the rpc does not reply with an error but with an
empty response `""` or `"0x"` which leads to `unpack errors` that make
the debugging process harder.

## Solution

- Detect `1.0.0` pools by proper error handling of the empty response
errors.



This PR introduces wETH transfers via a LockUnlock 1.0.0 token pool in
the integration test. While implementing this, we cleaned up a lot of
code that wasn't needed like manual price updates on the destination
chain. The commit plugin will update the prices. This PR improves the
coverage of our integration test by adding

- multiple token tx
- 1.0.0 token pools
- dest price updates through commitStore
- multiple priced assets
- lockUnlock token pools

---------

Co-authored-by: Rens Rooimans <github@rensrooimans.nl>