[lock/unlock] test coverage (#564)
## Motivation

Test coverage is wanting, so this PR adds a bunch of tests to improve
coverage. Found a bug in the process, although it not a major one.

## Solution

Add more tests. Also axed the caching logic for pending transfers for
now, we can get to it later when we nail down plugin biz logic.