Allow customizing `BackupPollerDelayBlock` or disabling Backup LogPoller entirely for tests (#11850)
* Refactor NewLogPoller() params into logpoller.Opts struct, add BackupPollerBlockDela

* Fix chain reader test & logpoller test

This failing chainreader test had FinalityDepth set to 0, so LogPoller was starting after
the block where the logs were emitted. It was only passing due to
Backup LogPoller.  With Backup LogPoller disabled, we have to increase
the finality depth so that LogPoller will start a few blocks back (at
the first unfinalized block).

The failing logpoller test was looking for an extra log entry emitted by BackupLogPoller.
Now we can run it with BackupLogPoller disabled (leading to a more robust test) and stop
looking for that

* Fix LogPoller_Replay test

* Address SonarQube CodeSmell: reduce cognitive complexity of lp.run()

* Add BackupLogPollerBlockDelay to toml config

* Fix some merge conflicts

* Restore KeepFinalizedBlocksDepth: 1000

* Use default BackupPollerBlockDelay: 0 instead of explicitly setting to 0

* disable backup log poller in lp smoke tests

* Fix generated docs

* Add missing web resolver testdata

---------

Co-authored-by: Bartek Tofel <bartek.tofel@smartcontract.com>