CCIP-1704 Checking for curse on Source Chain (#580)
## Motivation

CCIP doesn't verify if the source is cursed during OCR2 phases. We want
to add these check to make the system more robust.

## Solution

Additional check to verify if the source chain is cursed when checking
if CommitStore is down.
The first call to a specific OnRampReader requires an additional RPC
call to get staticConfig from the chain. StaticConfig is cached forever.
Next calls only get `ArmProxy` address from the cached config and ask
ARM on the source chain whether it's cursed or not.

This additional check is applied to both Commit and Exec, during
Observation phase