CCIP-1704 More robust checks of RMN/Chain every phase (#583)
More robust way of verifying if CCIP should halt processing, it's based
on four items:
1. Source chain is healthy (this is verified by checking if source
LogPoller saw finality violation)
2. Dest chain is healthy (this is verified by checking if destination
LogPoller saw finality violation)
3. CommitStore is down (this is verified by checking if CommitStore is
down and destination RMN is not cursed)
4. Source chain is cursed (this is verified by checking if source RMN is
not cursed)

Whenever any of the above checks fail, the chain is considered unhealthy
and the CCIP should stop
processing messages. Additionally, when the chain is unhealthy, this
information is considered "sticky"
and is cached for 30 minutes. This may lead to some false-positives, but
in this case, we want to be extra cautious and avoid executing any
reorged messages.

Health checks are now verified in every OCR2 phase and during
Observation and ShouldTransmit we enforce reading data from chain.
Additionally, to reduce the number of calls to the RPC, we cache RMN
curse state for 20 seconds.