[lock/unlock] OP L1 finalization in bridge adapter (#606)
## Motivation

The Optimism L1 bridge adapter needs to be able to both prove
withdrawals and finalize them on L1.

## Solution

* Add Optimism withdrawal proving and finalization on L1 functionality
to the bridge adapter. Tweak the scripts to prove and finalize
withdrawals through the adapters rather than on the Optimism contracts
directly to prove that they work.
* Fix a bug with ETH cross-chain transfers, on both L1 and L2 adapters:
since Optimism transfers ETH rather than WETH cross-chain through their
native bridge, stuff ends up reverting when we call `withdraw` on the
WETH contract prior to bridging.
* Fix a bug where we were using `withdrawTo` without actually sending
the ETH along with the call. The Optimism contracts dependency we're
using isn't bedrock so the `withdrawTo` function signature did not have
`payable`. Instead we copy/paste the `withdrawTo` function signature
directly from
[`L2StandardBridge.sol`](https://github.com/ethereum-optimism/optimism/blob/f707883038d527cbf1e9f8ea513fe33255deadbc/packages/contracts-bedrock/src/L2/L2StandardBridge.sol#L114-L122),
since we can't import it directly due to the exact 0.8.15 version
specified in the contract's pragma.

## Follow ups

* Direct follow up: support multiple-stage finalization in
Rebalancer.sol. Right now Rebalancer.sol assumes that a single
successful call to `finalizeWithdrawERC20` will give it access to funds,
however this is not the case for Optimism. This will be implemented in
an upcoming PR to keep diffs small and manageable.
* Future work: Explore the differences between `bridgeETHTo` and
`bridgeERC20To`, which are on `StandardBridge.sol`, vs. `depositERC20To`
and `withdrawTo` which are on L1StandardBridge.sol and
L2StandardBridge.sol respectively.
* Future work: Use contracts-bedrock imports instead of plain contracts.