CCIP-1718 optimize load tests to support more number of networks (#605)
## Motivation
With more number of networks and lanes ( 380 lanes , 20 chains) in
scalability tests, the memory consumption of the tests was exponentially
increasing

## Solution

- Got rid of unused elements from `CCIPLane` structs.
- Created `CCIPOptimizedLane` as a lighter version of `CCIPLane` struct
to be used in load test and used in loadgens.
- There are several contracts (for example token pools, price registry)
which are only used during lane/env set up , but not used further in
test. With OptimizeSpace method, set such unused contract references to
nil.