CCIP-1716 Adding retention to filters used by LogPoller (#530)
## Motivation

The goal of this PR is to reduce the number of logs and blocks we keep
in the database by utilizing a built-in retention mechanism in
LogPoller.
Requires paging for smooth deployment
https://github.com/smartcontractkit/chainlink/pull/12060

## Solution

This PR enables retention for all the LogPoller's filters registered by
CCIP. Additionally, to avoid pushing too much pressure during deletion
(especially the first run will have a lot of logs to remove) we've
updated `LogPrunePageSize` to 10k. Please see the original PR in the
chainlink repo to learn more about paging and its impact on the
database. `LogPrunePageSize` is altered in the `fallback.toml` to avoid
the necessity of setting this value for every chain that CCIP is
deployed on.