Finally found the bug!
There's a case where Keepers registers a filter with one name, then
unregisters it, then re-registers a filter with the same name. There was
an optimization where instead of computing the hash of the events &
topics for a filter in two different for loops, it saves it the first time
and looks it up in a map the second time. But the keys for the map were
the filter names. The first loop deals with removed filters and the
second with new filters. They both had the same filter name, but
different addresses & topics.  This might be a bug in Keepers, but I
can at least make this code more robust since it should be possible to
update a filter's topics & addresses but keep the name the same.