Use last_Processed.BlockNumber - finalityDepth when finality tags are disabled
This is preferable to reading last_Processed.FinalizedBlockNumber because it allows for
increasing the finalityDepth parameter to get the node unstuck after a finality violation.
Without this change, the only way to get unstuck is to delete blocks from the db

We can also use this as a backup if re-org detection fails (even when useFinalityTags=true).
If the presently configured finality depth is deeper than the depth according to what was
saved in the db, and there is a matching ancestor within that depth, we can recover by
updating some blocks which were thought to be finalized based on the configuration at the
time the last processed block was saved.