Remove error on missing value in destTokenPricesUSD (#692)
## Motivation
Self serve will not offer a full mapping of source to dest tokens. This
PR enables optimistic aggregate rate limiting, meaning it will only
calculate the rate limit value for tokens which are mapped. For all
lanes before 1.5, this means no change. For the 1.5+ lanes it means no
token will be included. This will be fixed in a future PR, before 1.5
will ever be deployed.