Add default pricing to onramp (#719)
The self serve PR introduced the option to not price tokens. This PR
expands upon that by introducing a default for pricing unconfigured
tokens.

- 	DefaultTokenFeeUSDCents
- 	DefaultTokenDestGasOverhead
- 	DefaultTokenDestBytesOverhead

This impacted the contract size, therefore the number of optimizations
had to be reduced to 4k, even after some golfing. While reviewing, it
was discovered the OffRamp could use the same number of optimizations as
all other CCIP code (26k) and no longer required the low 10k default.

Some misc things

- Inlined 2 single use modifiers in the onRamp
- Removed unused errors and events